get rid of compilation warnings

+++ example/170_oo_socket_koan/source01/src/ISocket.h
@@ -5,11 +5,12 @@
 #ifndef ISOCKET_H
 #define ISOCKET_H
 #include <string>
+#include <cstdint>
 
 class ISocket {
 public:
   virtual ~ISocket() = default;
-  virtual bool open(int port) = 0;
+  virtual bool open(uint16_t port) = 0;
   virtual void close() = 0;
   virtual bool send(const std::string &data) = 0;
   virtual std::string receive() = 0;
